Move types test into tests/types_test.rs
